Theories in communication and entropy of electrical signals in the brain of species.
Energy cannot be created or destroyed. Thoughts are electrical signals created or recieved in the brain. Electricity cannot be created according to rules of science. This theory investigates common regions in brain matter among many species to determine if entropy and laws of chemistry and physics apply to thoughts created in the mind.